refactor Car 클래스를 필드로 가지는 일급컬렉션 추가
