+++ src/main/java/racing/domain/Car.java
@@ -1,0 +1,18 @@
+package racing.domain;
+
+public class Car {
+    private int location;
+
+    public Car(int location) {
+        this.location = location;
+    }
+
+    public Car move(int distance) {
+        this.location += distance;
+        return new Car(this.location);
+    }
+
+    public int getLocation() {
+        return this.location;
+    }
+}
