+++ src/main/java/racing/domain/Cars.java
@@ -1,0 +1,30 @@
+package racing.domain;
+
+import racing.domain.Car;
+import racing.dto.GameResult;
+import racing.util.RandomInt;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class Cars { // 일급컬렉션
+    private List<Car> cars;
+
+    public Cars(List<Car> cars){
+        this.cars = cars;
+    }
+
+    public List<Car> getCars(){
+        return this.cars;
+    }
+
+    public GameResult playOnceTurn() {
+        List<Integer> onceResult = new ArrayList<>();
+
+        for(int i = 0; i < cars.size(); i++) {
+            this.cars.get(i).move(RandomInt.makeRandomInt());
+            onceResult.add(this.cars.get(i).getLocation());
+        }
+        return new GameResult(onceResult);
+    }
+}
